Initialize simple app: connect store, initialize simple routes, render basic App component
