+++ client/index.html
@@ -8,6 +8,7 @@
   <link rel="stylesheet" type="text/css" href="css/index.css">
 </head>
 <body>
+  <div id="app"></div>
   <div id="full-post">
     <span class="close-x right">&times;</span>
     <h2 class="subject-header"></h2>
