+++ client/js/components/app.js
@@ -1,0 +1,9 @@
+var React = require("react");
+
+var App = function(){
+  return(
+    <p>Rendering App!</p>
+  )
+};
+
+module.exports = App;
