+++ client/js/components/routes.js
@@ -1,0 +1,20 @@
+var React = require("react");
+var ReactDOM = require("react-dom");
+
+var router = require('react-router');
+var Router = router.Router;
+var Route = router.Route;
+var hashHistory = router.hashHistory;
+var IndexRoute = router.IndexRoute;
+var Link = router.Link;
+
+var App = require("./app");
+
+var routes = (
+  <Router history={hashHistory}>
+    <Route path="/" component={App}>
+    </Route>
+  </Router>
+);
+
+module.exports = routes;
