+++ client/js/reducers/reducer.js
@@ -1,0 +1,6 @@
+var React = require("react");
+
+var actions = require("../actions/actions")
+var reducer = function(){}
+
+module.exports = reducer;
