+++ package.json
@@ -97,6 +97,7 @@
     "react": "^15.3.1",
     "react-dom": "^15.3.1",
     "react-redux": "^4.4.5",
+    "react-router": "^3.0.0",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0"
   }
